requirements.txt
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==42.0.8
cssbeautifier==1.15.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.7
django-allauth==0.63.3
django-cors-headers==4.4.0
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.2
django-debug-toolbar==4.4.2
django-dynamic-breadcrumbs @ git+https://github.com/g0dnerd/django-dynamic-breadcrumbs.git@80fd2b923d67acf288f12f1d8db7891d388c0b5d
django-environ==0.11.2
django-extensions==3.2.3
django-google-auth==0.4.0
django-storages==1.14.3
django-termsandconditions @ git+https://github.com/g0dnerd/django-termsandconditions.git@main
djangorestframework==3.15.2
djlint==1.34.1
docopt==0.6.2
EditorConfig==0.12.4
executing==2.0.1
fastjsonschema==2.20.0
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.0
grpcio-status==1.65.0
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gettext==5.0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yarg==0.1.9
